Convert to TS and downcase TPEUBGSZ utilities
